Moved SuddenPlusValue data class to DataClasses file
