--- app/src/main/java/com/duraz0rz/suddenpluscalculator/SuddenPlusValue.kt
@@ -1,3 +1,0 @@
-package com.duraz0rz.suddenpluscalculator
-
-data class SuddenPlusValue(val highSpeed : String, val minWhiteNumber: Long, val maxWhiteNumber: Long? = null)
