+++ app/src/main/java/com/duraz0rz/suddenpluscalculator/dataClasses/DataClasses.kt
@@ -1,3 +1,3 @@
 package com.duraz0rz.suddenpluscalculator.dataClasses
 
-data class SuddenPlusTableParams(val minBPM : Int?, val maxBPM : Int? = null, val greenNumber : Int?)
+data class SuddenPlusValue(val highSpeed : String, val minWhiteNumber: Long, val maxWhiteNumber: Long? = null)
