+++ app/src/test/java/com/duraz0rz/suddenpluscalculator/SuddenPlusCalculatorTest.kt
@@ -1,5 +1,6 @@
 package com.duraz0rz.suddenpluscalculator
 
+import com.duraz0rz.suddenpluscalculator.dataClasses.SuddenPlusValue
 import com.natpryce.hamkrest.assertion.assertThat
 import com.natpryce.hamkrest.closeTo
 import com.natpryce.hamkrest.equalTo
@@ -42,20 +43,20 @@ class SuddenPlusCalculatorTest {
         val actualList = subject.generateSuddenPlusTable(bpm = 120, maxBpm = 160, greenNumber = 400)
 
         val expectedList = listOf(
-                SuddenPlusValue(highSpeed = "0.50", minWhiteNumber =  862, maxWhiteNumber = 816),
-                SuddenPlusValue(highSpeed = "0.75", minWhiteNumber =  793, maxWhiteNumber = 724),
-                SuddenPlusValue(highSpeed = "1.00", minWhiteNumber =  724, maxWhiteNumber = 632),
-                SuddenPlusValue(highSpeed = "1.25", minWhiteNumber =  655, maxWhiteNumber = 540),
-                SuddenPlusValue(highSpeed = "1.50", minWhiteNumber =  586, maxWhiteNumber = 448),
-                SuddenPlusValue(highSpeed = "1.75", minWhiteNumber =  517, maxWhiteNumber = 356),
-                SuddenPlusValue(highSpeed = "2.00", minWhiteNumber =  448, maxWhiteNumber = 264),
-                SuddenPlusValue(highSpeed = "2.25", minWhiteNumber =  379, maxWhiteNumber = 172),
-                SuddenPlusValue(highSpeed = "2.50", minWhiteNumber =  310, maxWhiteNumber = 80),
-                SuddenPlusValue(highSpeed = "2.75", minWhiteNumber =  241, maxWhiteNumber = -11),
-                SuddenPlusValue(highSpeed = "3.00", minWhiteNumber =  172, maxWhiteNumber = -103),
-                SuddenPlusValue(highSpeed = "3.25", minWhiteNumber =  103, maxWhiteNumber = -195),
-                SuddenPlusValue(highSpeed = "3.50", minWhiteNumber =   34, maxWhiteNumber = -287),
-                SuddenPlusValue(highSpeed = "3.75", minWhiteNumber =  -34, maxWhiteNumber = -379),
+                SuddenPlusValue(highSpeed = "0.50", minWhiteNumber = 862, maxWhiteNumber = 816),
+                SuddenPlusValue(highSpeed = "0.75", minWhiteNumber = 793, maxWhiteNumber = 724),
+                SuddenPlusValue(highSpeed = "1.00", minWhiteNumber = 724, maxWhiteNumber = 632),
+                SuddenPlusValue(highSpeed = "1.25", minWhiteNumber = 655, maxWhiteNumber = 540),
+                SuddenPlusValue(highSpeed = "1.50", minWhiteNumber = 586, maxWhiteNumber = 448),
+                SuddenPlusValue(highSpeed = "1.75", minWhiteNumber = 517, maxWhiteNumber = 356),
+                SuddenPlusValue(highSpeed = "2.00", minWhiteNumber = 448, maxWhiteNumber = 264),
+                SuddenPlusValue(highSpeed = "2.25", minWhiteNumber = 379, maxWhiteNumber = 172),
+                SuddenPlusValue(highSpeed = "2.50", minWhiteNumber = 310, maxWhiteNumber = 80),
+                SuddenPlusValue(highSpeed = "2.75", minWhiteNumber = 241, maxWhiteNumber = -11),
+                SuddenPlusValue(highSpeed = "3.00", minWhiteNumber = 172, maxWhiteNumber = -103),
+                SuddenPlusValue(highSpeed = "3.25", minWhiteNumber = 103, maxWhiteNumber = -195),
+                SuddenPlusValue(highSpeed = "3.50", minWhiteNumber = 34, maxWhiteNumber = -287),
+                SuddenPlusValue(highSpeed = "3.75", minWhiteNumber = -34, maxWhiteNumber = -379),
                 SuddenPlusValue(highSpeed = "4.00", minWhiteNumber = -103, maxWhiteNumber = -471)
         )
